# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.1"
  title: Musica
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /helloworld:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /gamesystems:
    x-swagger-router-controller: gamesystem.controller
    get:
      description: get the game system list
      operationId: getMusicas
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetGameSystemListResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse" 

    post:
      description: add a new game system to the list
      operationId: addMusicas
      parameters:
        - in: body
          name: gamesystem
          description: The Game System to create.
          schema:
            $ref: "#/definitions/GameSystem"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetGameSystemResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /gamesystems/{id}:
    x-swagger-router-controller: gamesystem.controller
    get:
      description: get a game system
      operationId: getMusicasbyId
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetGameSystemResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a game system
      operationId: updateMusicas
      # define the parameters
      parameters:
        - name: id
          description: game system id
          type: number
          in: path
          required: true
        - in: body
          name: gamesystem
          description: The Game System to update.
          schema:
            $ref: "#/definitions/GameSystem"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetGameSystemResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
    
    delete:
      description: delete a game system
      operationId: deleteMusica

      # define the parameters
      parameters:
        - name: id
          description: game system id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"    
   
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  GameSystem:
    type: object
    properties:
      Genero:
        type: string
        description: Genero of the Game System
      Autor:
        type: string
        description: Autor of the Game System
      Nombre:
        type: string
        description: Nombre of the Game System
      Duración:
        type: string
        description: Duracion of the Game System
      
    required:
      - Genero
      - Autor
      - Nombre
      - Duración
      
  GetGameSystemResponse:
    type: object
    properties:
      Genero:
        type: string
        description: Genero of the Game System
      Autor:
        type: string
        description: Autor of the Game System
      Nombre:
        type: string
        description: Nombre of the Game System
      Duración:
        type: string
        description: Duracion of the Game System
     
    required:
      - Genero
      - Autor
      - Nombre
      - Duración
     
 
  GetGameSystemListResponse:
    required:
      - gamesystems
    properties:
      gamesystems:
        type: array
        items:
          $ref: "#/definitions/GetGameSystemResponse"
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
     
      
 
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string